requirements.txt
scikit-learn~=0.24.1
numpy~=1.20.2
pandas~=1.2.4
opencv-python~=4.5.1.48
Pillow~=8.2.0
matplotlib~=3.4.1
seaborn~=0.11.1
psutil~=5.8.0
PyDrive~=1.3.1
py-cpuinfo~=8.0.0
streamlit~=0.80.0